requirements.txt

fastapi==0.109.0
uvicorn==0.23.2
PyYAML==6.0
pydantic>=2.3
openai==2.3.0
langchain>=0.0.348,<1.0.0
langchain-openai>=0.1.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint>=0.2.0
PyGithub>=2.1.1
GitPython>=3.1.0
psycopg2==2.9.11